refactor: move dashboard filters
