+++ resources/js/Pages/Dashboard.tsx
@@ -68,29 +68,12 @@ import { PERMISSION_ENUM } from '@/Support/Enums/permissionEnum';
 import { AttachmentStatusBarChartInterface, AttachmentStatusOfTrainsetResource, AttachmentStatusOfWorkstationResource, ReturnedProductStatusPieChartInterface } from '@/Support/Interfaces/Others';
 import ReturnedProductStatusPieChart from './Partials/ReturnedProductStatusPieChart';
 import { fetchEnumLabels } from '@/Helpers/enumHelper';
+import Filters from './Partials/Filters';
 
 export default function Dashboard({ auth, data }: PageProps) {
     const [localizedReturnedProductStatuses, setLocalizedReturnedProductStatuses] = useState<
         Record<string, string>
     >({});
-    const [open, setOpen] = useState(false);
-    const [yearFilterOpen, setYearFilterOpen] = useState(false);
-    const [monthFilterOpen, setMonthFilterOpen] = useState(false);
-    const years = [2022, 2023, 2024, 2025, 2026, 2027, 2028, 2029, 2030];
-    const months = [
-        { value: 1, label: 'Januari' },
-        { value: 2, label: 'Februari' },
-        { value: 3, label: 'Maret' },
-        { value: 4, label: 'April' },
-        { value: 5, label: 'Mei' },
-        { value: 6, label: 'Juni' },
-        { value: 7, label: 'Juli' },
-        { value: 8, label: 'Agustus' },
-        { value: 9, label: 'September' },
-        { value: 10, label: 'Oktober' },
-        { value: 11, label: 'November' },
-        { value: 12, label: 'Desember' },
-    ];
     const [openTrainset, setOpenTrainset] = useState(false);
     // const [value, setValue] = useState('');
     const [value, setValue] = useState(data['project'] !== null ? data['project'] : '');
@@ -316,7 +299,6 @@ export default function Dashboard({ auth, data }: PageProps) {
     }, []);
 
     useEffect(() => {
-        console.log('locale');
         fetchEnumLabels('ReturnedProductStatusEnum')
             .then(setLocalizedReturnedProductStatuses)
             .catch((error) => console.error('Failed to fetch localized statuses:', error));
@@ -344,213 +326,7 @@ export default function Dashboard({ auth, data }: PageProps) {
                                 </h2>
                             </div>
                             <div className="grid grid-cols-1 gap-2 md:grid-cols-3">
-                                <Popover open={open} onOpenChange={setOpen}>
-                                    <PopoverTrigger className=' ' asChild>
-                                        <Button
-                                            variant='outline'
-                                            role='combobox'
-                                            className='w-25 justify-between md:w-40'
-                                            aria-expanded={open}
-                                        >
-                                            {value
-                                                ? project.find(
-                                                    (projectItem) => projectItem.value === value,
-                                                )?.label
-                                                : t('pages.dashboard.index.select_project')}
-                                            <ChevronsUpDown className='ml-2 h-4 w-4 shrink-0 opacity-50' />
-                                        </Button>
-                                    </PopoverTrigger>
-                                    <PopoverContent className='w-[200px] p-0'>
-                                        <Command>
-                                            <CommandInput
-                                                placeholder={t('pages.dashboard.index.find_project')}
-                                            />
-                                            <CommandList>
-                                                <CommandEmpty>
-                                                    {t('pages.dashboard.index.project_not_found')}
-                                                </CommandEmpty>
-                                                <CommandGroup>
-                                                    {
-                                                        // @ts-ignore
-                                                        data['projectDetail'].map((projectItem) => (
-                                                            <Link
-                                                                key={projectItem.id}
-                                                                href={`/dashboard/${projectItem.id}`}
-                                                            >
-                                                                <CommandItem
-                                                                    value={`/dashboard/${projectItem.name}`}
-                                                                    onSelect={(currentValue) => {
-                                                                        setValue(
-                                                                            currentValue === value
-                                                                                ? ''
-                                                                                : currentValue,
-                                                                        );
-                                                                        setOpen(false);
-                                                                    }}
-                                                                    key={projectItem.value}
-                                                                >
-                                                                    <Check
-                                                                        className={cn(
-                                                                            'mr-2 h-4 w-4',
-                                                                            value === projectItem.name
-                                                                                ? 'opacity-100'
-                                                                                : 'opacity-0',
-                                                                        )}
-                                                                    />
-                                                                    {projectItem.name}
-                                                                </CommandItem>
-                                                            </Link>
-                                                        ))
-                                                    }
-                                                </CommandGroup>
-                                            </CommandList>
-                                        </Command>
-                                    </PopoverContent>
-                                </Popover>
-                                {checkPermission([PERMISSION_ENUM.RETURNED_PRODUCT_CREATE]) && (
-                                    <>
-                                        <Popover open={yearFilterOpen} onOpenChange={setYearFilterOpen}>
-                                            <PopoverTrigger asChild>
-                                                <Button
-                                                    variant='outline'
-                                                    role='combobox'
-                                                    className='w-25 justify-between md:w-40'
-                                                    aria-expanded={yearFilterOpen}
-                                                >
-                                                    {dateFilterValue.year
-                                                        ? dateFilterValue.year
-                                                        : "Filter Tahun"}
-                                                    <ChevronsUpDown className='ml-2 h-4 w-4 shrink-0 opacity-50' />
-                                                </Button>
-                                            </PopoverTrigger>
-                                            <PopoverContent className='w-[200px] p-0'>
-                                                <Command>
-                                                    <CommandList>
-                                                        <CommandEmpty>
-                                                            {"Year Not Found"}
-                                                        </CommandEmpty>
-                                                        <CommandGroup>
-                                                            {dateFilterValue.year && (
-                                                                <CommandItem onSelect={() => {
-                                                                    setDateFilterValue({
-                                                                        year: '',
-                                                                        month: 0
-                                                                    });
-                                                                    setYearFilterOpen(false);
-                                                                }}>
-                                                                    {t('components.generic_data_selector.actions.clear_selection')}
-                                                                </CommandItem>
-                                                            )}
-                                                            {years.map((yearItem) => (
-                                                                <CommandItem
-                                                                    value={yearItem.toString()}
-                                                                    onSelect={(currentValue) => {
-                                                                        setDateFilterValue({
-                                                                            year: currentValue === dateFilterValue.year
-                                                                                ? ''
-                                                                                : currentValue.toString(),
-                                                                            month: dateFilterValue.month
-                                                                        });
-                                                                        setYearFilterOpen(false);
-                                                                    }}
-                                                                    key={yearItem}
-                                                                >
-                                                                    <Check
-                                                                        className={cn(
-                                                                            'mr-2 h-4 w-4',
-                                                                            dateFilterValue.year === yearItem.toString()
-                                                                                ? 'opacity-100'
-                                                                                : 'opacity-0',
-                                                                        )}
-                                                                    />
-                                                                    {yearItem}
-                                                                </CommandItem>
-                                                            ))}
-                                                        </CommandGroup>
-                                                    </CommandList>
-                                                </Command>
-                                            </PopoverContent>
-                                        </Popover>
-                                        
-                                        {/* <GenericDataSelector
-                                            // TODO: redesain dis shts🗿
-                                            id="year-filter"
-                                            data={years.map((year, index) => ({id: index, value: year}))}
-                                            setSelectedData={id => {
-                                                setYearFilterValue(id ? years[id].toString() : '');
-                                            }}
-                                            selectedDataId={yearFilterValue ? years.indexOf(parseInt(yearFilterValue)) : null}
-                                            placeholder={'Choose'}
-                                            renderItem={item =>
-                                                `${item.value}`
-                                            }
-                                            nullable
-                                        /> */}
-                                        {dateFilterValue.year && (
-                                            <Popover open={monthFilterOpen} onOpenChange={setMonthFilterOpen}>
-                                                <PopoverTrigger asChild>
-                                                    <Button
-                                                        variant='outline'
-                                                        role='combobox'
-                                                        className='w-25 justify-between md:w-40'
-                                                        aria-expanded={monthFilterOpen}
-                                                    >
-                                                        {dateFilterValue.month
-                                                            ? months.find(month => month.value === dateFilterValue.month)?.label
-                                                            : "Filter Bulan"}
-                                                        <ChevronsUpDown className='ml-2 h-4 w-4 shrink-0 opacity-50' />
-                                                    </Button>
-                                                </PopoverTrigger>
-                                                <PopoverContent className='w-[200px] p-0'>
-                                                    <Command>
-                                                        <CommandList>
-                                                            <CommandEmpty>
-                                                                {"Month Not Found"}
-                                                            </CommandEmpty>
-                                                            <CommandGroup>
-                                                                {dateFilterValue.month !== 0 && (
-                                                                    <CommandItem onSelect={() => {
-                                                                        setDateFilterValue({
-                                                                            year: dateFilterValue.year,
-                                                                            month: 0,
-                                                                        })
-                                                                        setMonthFilterOpen(false)
-                                                                    }}>
-                                                                        {t('components.generic_data_selector.actions.clear_selection')}
-                                                                    </CommandItem>
-                                                                )}
-                                                                {months.map((monthItem) => (
-                                                                    <CommandItem
-                                                                        value={monthItem.label}
-                                                                        onSelect={() => {
-                                                                            setDateFilterValue({
-                                                                                year: dateFilterValue.year,
-                                                                                month: monthItem.value,
-                                                                            });
-                                                                            setMonthFilterOpen(false);
-                                                                        }}
-                                                                        key={monthItem.label}
-                                                                    >
-                                                                        <Check
-                                                                            className={cn(
-                                                                                'mr-2 h-4 w-4',
-                                                                                dateFilterValue.month === monthItem.value
-                                                                                    ? 'opacity-100'
-                                                                                    : 'opacity-0',
-                                                                            )}
-                                                                        />
-                                                                        {monthItem.label}
-                                                                    </CommandItem>
-                                                                ))}
-                                                            </CommandGroup>
-                                                        </CommandList>
-                                                    </Command>
-                                                </PopoverContent>
-                                            </Popover>
-                                        )}
-                                    </>
-                                    
-                                )}
+                                <Filters data={data} dateFilterValue={dateFilterValue} setDateFilterValue={setDateFilterValue}/>
                             </div>
                         </div>
